refactor data-permissions selectors
