+++ frontend/src/metabase-types/store/admin.ts
@@ -1,5 +1,11 @@
+import { GroupsPermissions } from "metabase-types/types/Permissions";
+
 export interface AdminState {
   app: AdminAppState;
+  permissions: {
+    dataPermissions: GroupsPermissions;
+    originalDataPermissions: GroupsPermissions;
+  };
 }
 
 export interface AdminAppState {
