+++ frontend/src/metabase-types/types/Metadata.ts
@@ -12,11 +12,24 @@ export type Metadata = {
   fields: { [id: FieldId]: FieldMetadata };
   metrics: { [id: MetricId]: MetricMetadata };
   segments: { [id: SegmentId]: SegmentMetadata };
+
+  database: (id: number) => DatabaseMetadata;
 };
 
 export type DatabaseMetadata = Database & {
   tables: TableMetadata[];
   tables_lookup: { [id: TableId]: TableMetadata };
+
+  schema: (name: string | null) => SchemaMetadata;
+  getSchemas: () => SchemaMetadata[];
+};
+
+export type SchemaMetadata = {
+  id: string;
+  name: string;
+  tables: TableMetadata[];
+
+  getTables: () => TableMetadata[];
 };
 
 export type TableMetadata = Table & {
