+++ frontend/src/metabase-types/types/Table.ts
@@ -16,6 +16,7 @@ export type Table = {
   db_id: DatabaseId;
 
   schema?: SchemaName;
+  schema_name: SchemaName;
   name: string;
   display_name: string;
