+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/data-permissions.unit.spec.js
@@ -1,7 +1,4 @@
-import {
-  getDataFocusSidebar,
-  getGroupsDataPermissionEditor,
-} from "./data-permissions";
+import { getDataFocusSidebar, getGroupsDataPermissionEditor } from ".";
 import { normalizedMetadata } from "./data-permissions.unit.spec.fixtures";
 
 const initialPermissions = {
