+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/index.ts
@@ -1,0 +1,8 @@
+export * from "./breadcrumbs";
+export * from "./data-sidebar";
+export * from "./diff";
+export * from "./fields";
+export * from "./group-sidebar";
+export * from "./permission-editor";
+export * from "./schemas";
+export * from "./tables";
