+++ frontend/src/metabase/admin/permissions/selectors/types.ts
@@ -1,0 +1,23 @@
+export type GroupRouteParams = {
+  groupId: number;
+  databaseId?: number;
+  schemaName?: string;
+};
+
+export type RawGroupRouteParams = {
+  groupId: string;
+  databaseId: string;
+  schemaName?: string;
+};
+
+export type DataRouteParams = {
+  databaseId: number;
+  schemaName?: string;
+  tableId?: number;
+};
+
+export type RawDataRouteParams = {
+  databaseId: string;
+  schemaName?: string;
+  tableId?: string;
+};
